slow box, normal mis

+++ AABB.cpp
@@ -21,27 +21,45 @@ AABB AABB::merge(AABB box, Vec3 p) {
 	return { new_min, new_max };
 }
 
-double AABB::get_middle(int axis) {
+float AABB::get_middle(int axis) {
 	return 0.5 * (p_min_[axis] + p_max_[axis]);
 }
 
 bool AABB::intersect(Ray& ray) {
-	Vec3 inv_dir{ 1.0 / ray.direction.x(), 1.0 / ray.direction.y(), 1.0 / ray.direction.z() };
+	Vec3 inv_dir{ 1.0f / ray.direction.x(), 1.0f / ray.direction.y(), 1.0f / ray.direction.z() };
 
-	//double t_x_1 = (p_min_.x() - ray.origin.x()) * inv_dir.x();
-	//double t_x_2 = (p_max_.x() - ray.origin.x()) * inv_dir.x();
+	//float t_x_1 = (p_min_.x() - ray.origin.x()) * inv_dir.x();
+	//float t_x_2 = (p_max_.x() - ray.origin.x()) * inv_dir.x();
 
-	//double t_y_1 = (p_min_.y() - ray.origin.y()) * inv_dir.y();
-	//double t_y_2 = (p_max_.y() - ray.origin.y()) * inv_dir.y();
+	//float t_y_1 = (p_min_.y() - ray.origin.y()) * inv_dir.y();
+	//float t_y_2 = (p_max_.y() - ray.origin.y()) * inv_dir.y();
 
-	//double t_z_1 = (p_min_.z() - ray.origin.z()) * inv_dir.z();
-	//double t_z_2 = (p_max_.z() - ray.origin.z()) * inv_dir.z();
+	//float t_z_1 = (p_min_.z() - ray.origin.z()) * inv_dir.z();
+	//float t_z_2 = (p_max_.z() - ray.origin.z()) * inv_dir.z();
+	//float t_max = std::numeric_limits<float>::max();
+	//float t_min = 0.0;
+	//for (int i = 0; i < 3; i++)
+	//{
+	//	float t_enter = (p_min_[i] - ray.origin[i]) * inv_dir[i];
+	//	float t_exit = (p_max_[i] - ray.origin[i]) * inv_dir[i];
+	//	if (inv_dir[i] < 0)
+	//	{
+	//		std::swap(t_enter, t_exit);
+	//	}
+	//	t_max = std::min(t_exit, t_max);
+	//	t_min = std::max(t_enter, t_min);
+	//	if (t_max < t_min)
+	//	{
+	//		return false;
+	//	}
+	//}
+	//return t_max > eps;
 
 	Vec3 t1 = (p_min_ - ray.origin) * inv_dir;
 	Vec3 t2 = (p_max_ - ray.origin) * inv_dir;
 
-	double t_enter = std::max(std::max(std::min(t1.x(), t2.x()), std::min(t1.y(), t2.y())), std::min(t1.z(), t2.z()));
-	double t_exit = std::min(std::min(std::max(t1.x(), t2.x()), std::max(t1.y(), t2.y())), std::max(t1.z(), t2.z()));
+	float t_enter = std::max(std::max(std::min(t1.x(), t2.x()), std::min(t1.y(), t2.y())), std::min(t1.z(), t2.z()));
+	float t_exit = std::min(std::min(std::max(t1.x(), t2.x()), std::max(t1.y(), t2.y())), std::max(t1.z(), t2.z()));
 
 	if (t_enter <= t_exit && t_exit > eps)
 	{
