+++ Camera.h
@@ -4,7 +4,7 @@
 
 class Camera {
 public:
-	Camera(Vec3 eye, Vec3 lookat, Vec3 up, double fovy, int width, int height);
+	Camera(Vec3 eye, Vec3 lookat, Vec3 up, float fovy, int width, int height);
 	
 	Ray get_ray(int x, int y);
 	Ray sample_ray(int x, int y);
