+++ Texture.h
@@ -7,7 +7,7 @@ class Texture {
 public:
 	Texture() {}
 	Texture(std::string& path);
-	Vec3 get_color(double u, double v);
+	Vec3 get_color(float u, float v);
 
 	int width, height, channel;
 	float* data_;
